refactor : IDE-37 StorageManager 에러 추가 및 클래스 위치 수정
