+++ src/main/java/goorm/dbjj/ide/storageManager/model/ResourceType.java
@@ -1,4 +1,4 @@
-package goorm.dbjj.ide.model.type;
+package goorm.dbjj.ide.storageManager.model;
 
 public enum ResourceType {
     FILE, DIRECTORY;
